Separate app routes components in different folders and add Home.js component to home route
