+++ src/components/calculator/Button.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as styles from '../styles/components.module.css';
+import * as styles from '../../styles/components.module.css';
 
 function Button(props) {
   const { name, handleClick } = props;
